rename tests/utils.py to tests/testutils.py
