+++ tests/datasets/test_image_transforms.py
@@ -34,7 +34,7 @@ from lerobot.scripts.visualize_image_transforms import (
     save_each_transform,
 )
 from tests.artifacts.image_transforms.save_image_transforms_to_safetensors import ARTIFACT_DIR
-from tests.utils import require_x86_64_kernel
+from tests.testutils import require_x86_64_kernel
 
 
 @pytest.fixture
