+++ tests/envs/test_envs.py
@@ -23,7 +23,7 @@ from gymnasium.utils.env_checker import check_env
 import lerobot
 from lerobot.common.envs.factory import make_env, make_env_config
 from lerobot.common.envs.utils import preprocess_observation
-from tests.utils import require_env
+from tests.testutils import require_env
 
 OBS_TYPES = ["state", "pixels", "pixels_agent_pos"]
