+++ tests/motors/test_motors.py
@@ -45,7 +45,7 @@ import pytest
 
 from lerobot.common.robot_devices.utils import RobotDeviceAlreadyConnectedError, RobotDeviceNotConnectedError
 from lerobot.scripts.find_motors_bus_port import find_port
-from tests.utils import TEST_MOTOR_TYPES, make_motors_bus, require_motor
+from tests.testutils import TEST_MOTOR_TYPES, make_motors_bus, require_motor
 
 
 @pytest.mark.parametrize("motor_type, mock", TEST_MOTOR_TYPES)
