+++ tests/test_available.py
@@ -23,7 +23,7 @@ from lerobot.common.policies.act.modeling_act import ACTPolicy
 from lerobot.common.policies.diffusion.modeling_diffusion import DiffusionPolicy
 from lerobot.common.policies.tdmpc.modeling_tdmpc import TDMPCPolicy
 from lerobot.common.policies.vqbet.modeling_vqbet import VQBeTPolicy
-from tests.utils import require_env
+from tests.testutils import require_env
 
 
 @pytest.mark.parametrize("env_name, task_name", lerobot.env_task_pairs)
